Notifications, Home icons
Now it asks for the user permission to receive notifications, something that we have to explore, and it needs now, in the preferences, a button to turn the notifications on or off!
Home icons: changed the heart icon to a thumbs up, fits the app, because thumbs up for a carpool.
windowSearch its not working, why? Don't know yet, its doing the same as the home, when it comes to loading the data, but the button where the modal should appear its giving errors, something that needs a fix, going to look into it later.